# swagger.yaml
openapi: 3.0.3
info:
  title: QueryTube API
  description: |
    AI-powered video Q&A system with semantic search capabilities.
    Transform any video into an intelligent chatbot.
  version: 1.0.0
  contact:
    name: Vivek Kumar Singh
    url: https://github.com/reachvivek
  license:
    name: Custom Proprietary License
    url: https://github.com/reachvivek/QueryTube/blob/main/LICENSE.md
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://querytube.app/api
    description: Production server
...
paths:
...
  /embeddings:
    post:
      tags:
        - Embeddings
      summary: Generate vector embeddings
      description: Generates semantic embeddings using Mistral AI or OpenAI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - chunks
              properties:
                chunks:
                  type: array
                  items:
                    $ref: '#/components/schemas/Chunk'
                  example:
                    - text: Example transcript text
                      startTime: 0
                      endTime: 90
                model:
                  type: string
                  example: mistral-embed
                  enum:
                    - mistral-embed
                    - text-embedding-3-small
                provider:
                  type: string
                  example: mistral
                  enum:
                    - mistral
                    - openai
      responses:
        '200':
          description: Embeddings generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmbeddingsResponse'
  /upload-vectors:
    post:
      tags:
        - Embeddings
      summary: Upload vectors to Pinecone
      description: Uploads embeddings to Pinecone and generates AI summary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - videoId
                - chunks
                - embeddings
              properties:
                videoId:
                  type: string
                  format: uuid
                chunks:
                  type: array
                  items:
                    $ref: '#/components/schemas/Chunk'
                embeddings:
                  type: array
                  items:
                    type: array
                    items:
                      type: number
                embeddingProvider:
                  type: string
                  example: mistral
                metadata:
                  type: object
                  properties:
                    title:
                      type: string
                    youtubeUrl:
                      type: string
                    language:
                      type: string
                    uploader:
                      type: string
      responses:
        '200':
          description: Vectors uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  vectorsUploaded:
                    type: integer
                    example: 30
                  summary:
                    type: string
                    example: This video explains quantum mechanics fundamentals.
  /generate-summary:
    post:
      tags:
        - Transcript
      summary: Generate AI summary
      description: Generates AI-powered summary of video content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - videoId
              properties:
                videoId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Summary generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  summary:
                    type: string
                    example: This video explains the fundamentals of machine learning.
                  fromCache:
                    type: boolean
                    example: false
  /qa:
    post:
      tags:
        - Q&A
      summary: Ask question about video
      description: RAG-based question answering with semantic search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionRequest'
      responses:
        '200':
          description: Answer generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnswerResponse'
...
components:
...
  schemas:
...
    Chunk:
      type: object
      properties:
        chunkIndex:
          type: integer
          example: 0
        text:
          type: string
          example: In this video, we'll explore quantum mechanics...
        startTime:
          type: integer
          example: 0
          description: Start time in seconds
        endTime:
          type: integer
          example: 90
          description: End time in seconds
...
    EmbeddingsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        embeddings:
          type: array
          items:
            type: array
            items:
              type: number
        model:
          type: string
          example: mistral-embed
        provider:
          type: string
          example: mistral
        dimensions:
          type: integer
          example: 1024
        totalChunks:
          type: integer
          example: 30
    QuestionRequest:
      type: object
      required:
        - question
      properties:
        question:
          type: string
          example: What is quantum entanglement?
        videoId:
          type: string
          format: uuid
          nullable: true
          description: Filter to specific video, null for all videos
        provider:
          type: string
          enum:
            - groq
            - openai
            - claude
          example: groq
        model:
          type: string
          example: llama-3.3-70b-versatile
        systemPrompt:
          type: string
          nullable: true
        topK:
          type: integer
          example: 5
          minimum: 1
          maximum: 20
    AnswerResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        answer:
          type: string
          example: Quantum entanglement is a phenomenon where...
        chunks:
          type: array
          items:
            type: object
            properties:
              text:
                type: string
              score:
                type: number
                format: float
                example: 0.89
              timestamp:
                type: string
                example: "12:34"
              title:
                type: string
              videoId:
                type: string
        responseTime:
          type: number
          format: float
          example: 1.2
        model:
          type: string
          example: llama-3.3-70b-versatile
...